Add initial dumpling web launcher version
I opted to not reinvent everything and use existing projects, but it does make it a little janky in terms of code structure. The readme instructions should work if you wanted to change things, but this really shouldn't be used as a replacement for the existing architecture. I added credits on the page so that hopefully each project still gets their due.